+++ services/workflows-service/src/workflow-defintion/demo-workflow/index.ts
@@ -1,0 +1,5 @@
+import { upsertDemoEuKybFlow } from './eu/create-demo-workflow-eu';
+import { upsertDemoUsKybFlow } from './us/create-demo-workflow-us';
+import { createDemoWorkflow } from './create-demo-workflow';
+
+export { upsertDemoEuKybFlow, upsertDemoUsKybFlow, createDemoWorkflow };